requirements.txt
Flask==2.3.3
pytesseract==0.3.10
Pillow==9.5.0
pandas==2.1.1
scikit-learn==1.3.1
Werkzeug==2.3.7
